Support TeX/LaTeX.
In TeX semicolons are not used to terminate statements.
Also lists are uncommon so there is no need for the comma either.
But there is a another use case for this plugin.
It is a frequently task to append a '%' at the end of a line to avoid an
additional space being added by TeX.

Example:

```TeX
no-sp%
ace
```

which is typeset as "no-space". Without the '%' it would be 'no-sp ace'.