Fix block formatted arrays
When creating an array similar to:

```vim
$array = [
    'key' => 'item',
    'key2' => 'item2',
];
```
The plugin would put a semicolon on the last item in the array.

This commit fixes this and instead puts a comma at the end. While this
will leave a comma at the end of the last item, this is preferable to a
semicolon, and allows for a user to quickly build an array without
having to have a comma on the line below the item they are inserting.