Added CommaOrSemiColonPrevious Command
Added a command that will run cosco on the line directly above the
current line. This is usefull for creating a mapping that will check
the line of code you wrote after you hit enter to create a new line.

I have described how to use this command and provided an example
mapping for the described auto check in the readme.