Bugfix: For the following case:     func1(        arg1,        func2(),        |     )  ^      Cursor
Cosco wouldn't go to the last elseif condition, since it only looks, if there's only a semicolon after a round brace.
In this case we've a comma so a short fix of the regex from ');$' to ')[,;]$' should fix this issue.

To sum it up: Cosco shouldn't add a semicolon after 'func2(),' (result would be 'func2(),;') now.