Bugfix: Following example:      );  []
Cosco would add a comma after the previous line due to the square brackets. Should be fixed now
with the new regex pattern.

Also moved the condition if there's already a semicolon/comma to the bottom. But still need to think about the
condition steps so it's only temporary the last condition.