Bugfixes: 1. Case:
When the user creates a struct like this:

    struct Structure
    {

then Cosco would add a comma/semicolon after 'Structure'. There's now a condition in the
cosco_eval#ShouldSkip() function, but it'll be probably ported to the 'cosco_eval#Specials()' function.

2. Case:

When the user creates a function like that:

    Test *
    func_name(args)
    {

then Cosco would add a comma/semicolon after the '*', if the user types something in the next line of the star.
This should be fixed now by the new condition in the 'cosco_eval#Specials()' function.